% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{The clinical trial study data set}
\description{
Key variables include:
}
\details{
`ATRT` the treatment allocation;
`DTH` the death indicator, 1 means death, 0 otherwise;
`AGE` patient age;
`SEX` patient sex;
`RACE` patient race;
}
\keyword{data}
