% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{cleaning}
\alias{cleaning}
\title{Categorize age into age groups}
\usage{
cleaning(df)
}
\arguments{
\item{df}{the data set with age variable}
}
\value{
a new data set with categorized age variables
}
\description{
Categorize age into age groups
}
\examples{
data('adsl')
df = cleaning(df)
}
