% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{logit_test}
\alias{logit_test}
\title{Conduct logistic regression for certain outcome and predictors}
\usage{
logit_test(outcome, variables, df)
}
\arguments{
\item{outcome}{the target outcome variable}

\item{variables}{the predictor variables used in the logistic regression}

\item{df}{the dataset used in the analysis}
}
\description{
Conduct logistic regression for certain outcome and predictors
}
\examples{
data('adsl')
df = cleaning(df)
logit_test('DTH', 'ATRT', df)
}
