% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_strat}
\alias{plot_strat}
\title{Plot death outcome by treatment group stratified by a variable}
\usage{
plot_strat(x, col)
}
\arguments{
\item{x}{the target data frame}

\item{col}{the variable that the plot stratified on}
}
\value{
a bar plot faceted by axis
}
\description{
Plot death outcome by treatment group stratified by a variable
}
\examples{
data('adsl')
plot_strat(df, 'SEX')
}
